fix: add missing worker API endpoints for content generation and KB health
The frontend API client was written for the old ASP.NET backend, but many
worker routes were missing or had URL mismatches, causing 404/405 errors:

Content routes:
- POST /generate — AI content generation from random seed note
- POST /generate/from-note/:id — generation from a specific note
- POST /generations/:id/regenerate — full regeneration
- POST /generations/:id/regenerate/:medium — per-medium regeneration
- PUT /pieces/:id/approve, /reject — piece review actions
- GET /pieces/:id/export — markdown export
- PUT /pieces/:id/description, /tags — metadata updates
- POST /pieces/:id/send-to-draft — stub (422 until configured)
- Fix skip/take pagination (frontend) alongside page/pageSize
- Parse JSON tags columns and normalize nullable fields for frontend
- Fix count(*)::int PostgreSQL syntax → count(*) for D1/SQLite

KB Health routes:
- GET /orphan/:noteId/suggestions — Vectorize similarity search
- POST /orphan/:orphanId/link — add wiki-link (RESTful URL)
- POST /missing-embeddings/:noteId/requeue — re-queue embedding
- POST /large-notes/:noteId/summarize — AI summarization
- POST /large-notes/:noteId/split-suggestions — AI split analysis
- POST /large-notes/:noteId/apply-split — create split notes

+++ src/worker/src/routes/tags.ts
@@ -11,7 +11,7 @@ router.get('/', async (c) => {
   const rows = await db
     .select({
       tag: noteTags.tag,
-      count: sql<number>`count(*)::int`,
+      count: sql<number>`count(*)`,
     })
     .from(noteTags)
     .groupBy(noteTags.tag)
